refactor: split the sidebar chip family out of row-view
row-view.ts crossed the size cap when the review chip landed, along a seam the
golden already used: `buildSidebarRowView` is a priority ladder over live
engine/daemon/clock state, while the four chips take one stored field each and
answer with one cell. row-chips.ts holds all four together, because they share
the row's cell group and no glyph may mean two things.
